데미지 인디케이터, 진동 추가, 물 평타

+++ Game_Jam/Assets/01. Scripts/Objects/MoveObj/Torch.cs
@@ -78,7 +78,7 @@ public class Torch : MoveObj
 
     private IEnumerator GiveDamage()
     {
-        float damage = 0f;
+        int damage = 0;
 
         while (true)
         {
